Fix delete vehicle

+++ app/src/main/java/com/kd8bny/maintenanceman/classes/data/VehicleLogDBHelper.java
@@ -46,15 +46,14 @@ public class VehicleLogDBHelper extends SQLiteOpenHelper{
         db.execSQL("CREATE TABLE "
                 + TABLE_VEHICLE
                 + "("
-                + COLUMN_ID + " integer primary key autoincrement, "
-                + COLUMN_VEHICLE_REFID + " text not null, "
-                + COLUMN_VEHICLE_DATE + " text not null, "
-                + COLUMN_VEHICLE_ODO + " text not null, "
-                + COLUMN_VEHICLE_EVENT + " text not null, "
-                + COLUMN_VEHICLE_PRICE + " text not null, "
-                + COLUMN_VEHICLE_COMMENT + " text not null,"
-                + COLUMN_ICON + " text not null default '0'"
-                + ");");
+                + COLUMN_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, "
+                + COLUMN_VEHICLE_REFID + " TEXT NOT NULL, "
+                + COLUMN_VEHICLE_DATE + " TEXT NOT NULL, "
+                + COLUMN_VEHICLE_ODO + " TEXT NOT NULL, "
+                + COLUMN_VEHICLE_EVENT + " TEXT NOT NULL, "
+                + COLUMN_VEHICLE_PRICE + " TEXT NOT NULL, "
+                + COLUMN_VEHICLE_COMMENT + " TEXT NOT NULL,"
+                + COLUMN_ICON + " TEXT NOT NULL DEFAULT '0');");
     }
 
     public void insertEntry(String refID, HashMap<String, String> dataSet) {
@@ -62,7 +61,6 @@ public class VehicleLogDBHelper extends SQLiteOpenHelper{
         *String date, String odo, String task, String price, String comment, String icon
         */
         SQLiteDatabase db = getWritableDatabase();
-
         ContentValues contentValues = new ContentValues();
         contentValues.put(COLUMN_VEHICLE_REFID, refID);
         contentValues.put(COLUMN_VEHICLE_DATE, dataSet.get("Date"));
@@ -84,7 +82,7 @@ public class VehicleLogDBHelper extends SQLiteOpenHelper{
 
     public ArrayList<ArrayList> getFullVehicleEntries(String refID) {
         SQLiteDatabase db = getReadableDatabase();
-        String QUERY = String.format("SELECT * FROM %s WHERE %s = '%s';", TABLE_VEHICLE, COLUMN_VEHICLE_REFID, refID);
+        String QUERY = String.format("SELECT rowid, * FROM %s WHERE %s = '%s';", TABLE_VEHICLE, COLUMN_VEHICLE_REFID, refID);
         Cursor cursor = db.rawQuery(QUERY, null);
 
         ArrayList<ArrayList> vehicleList = new ArrayList<>();
@@ -93,7 +91,6 @@ public class VehicleLogDBHelper extends SQLiteOpenHelper{
             if (cursor.moveToFirst()) {
                 do {
                     ArrayList<String> temp = new ArrayList<>();
-
                     temp.add(cursor.getString(cursor.getColumnIndex(COLUMN_ICON)));
                     temp.add(cursor.getString(cursor.getColumnIndex(COLUMN_VEHICLE_DATE)));
                     temp.add(cursor.getString(cursor.getColumnIndex(COLUMN_VEHICLE_ODO)));
@@ -149,24 +146,23 @@ public class VehicleLogDBHelper extends SQLiteOpenHelper{
         return vehicleList;
     }
 
-    public void deleteEntry(ArrayList<String> arrayList){//TODO test
+    public void deleteEntry(ArrayList<String> arrayList){
         /* Data incoming format
         *String icon, String date, String odo, String event, String price, String comment
         */
         SQLiteDatabase db = getWritableDatabase();
-        String QUERY = String.format("DELETE FROM %s WHERE %s = %s " +
-                "AND %s = %s AND %s = %s AND %s = %s AND %s = %s;",
+        String QUERY = String.format("DELETE FROM %s WHERE %s = '%s' AND %s = '%s';",
                 TABLE_VEHICLE,
                 COLUMN_VEHICLE_DATE, arrayList.get(1),
-                COLUMN_VEHICLE_EVENT, arrayList.get(3),
-                COLUMN_VEHICLE_ODO, arrayList.get(2),
-                COLUMN_VEHICLE_PRICE, arrayList.get(4),
-                COLUMN_VEHICLE_COMMENT, arrayList.get(5));
+                COLUMN_VEHICLE_EVENT, arrayList.get(3));
         try {
-            db.rawQuery(QUERY, null);
+            db.beginTransaction();
+            db.execSQL(QUERY);
+            db.setTransactionSuccessful();
         }catch (Exception e) {
             e.printStackTrace();
         }finally {
+            db.endTransaction();
             db.close();
         }
     }
@@ -192,21 +188,76 @@ public class VehicleLogDBHelper extends SQLiteOpenHelper{
                         + newVersion + ", which will destroy all old data");
         db.beginTransaction();
         String TABLE_VEHICLE_NEW = TABLE_VEHICLE + "_new";
-        try {
-            Log.i(TAG, "trying");
-            db.execSQL("CREATE TABLE IF NOT EXISTS " + TABLE_VEHICLE_NEW + " AS SELECT * FROM " + TABLE_VEHICLE);
-            db.execSQL("ALTER TABLE " + TABLE_VEHICLE_NEW + " ADD COLUMN " + COLUMN_ICON + " text not null default '0'");
-            db.execSQL("DROP TABLE IF EXISTS " + TABLE_VEHICLE);
-            db.execSQL("ALTER TABLE " + TABLE_VEHICLE_NEW + " RENAME TO " + TABLE_VEHICLE);
-            onCreate(db);
-            db.setTransactionSuccessful();
-            Log.i(TAG, "success");
-        } catch (Exception e){
-            Log.e(TAG,e.toString());
-            Log.e(TAG, "Error updating db");
-        }finally {
-            db.setVersion(DB_VERSION);
-            db.endTransaction();
+        if (oldVersion < 3) {
+            try {
+                Log.i(TAG, "trying");
+                db.execSQL("CREATE TABLE IF NOT EXISTS " + TABLE_VEHICLE_NEW + " AS SELECT * FROM " + TABLE_VEHICLE);
+                db.execSQL("ALTER TABLE " + TABLE_VEHICLE_NEW + " ADD COLUMN " + COLUMN_ICON + " text not null default '0'");
+                db.execSQL("DROP TABLE IF EXISTS " + TABLE_VEHICLE);
+                db.execSQL("ALTER TABLE " + TABLE_VEHICLE_NEW + " RENAME TO " + TABLE_VEHICLE);
+                db.setTransactionSuccessful();
+                Log.i(TAG, "success");
+            } catch (Exception e) {
+                Log.e(TAG, e.toString());
+                Log.e(TAG, "Error updating db");
+            } finally {
+                //db.setVersion(DB_VERSION);
+                db.endTransaction();
+            }
         }
+
+        /*if (oldVersion == 3){
+            Log.i(TAG, "trying");
+            db.execSQL("CREATE TABLE "
+                    + TABLE_VEHICLE_NEW
+                    + "("
+                    + COLUMN_ID + " integer primary key autoincrement, "
+                    + COLUMN_VEHICLE_REFID + " text not null, "
+                    + COLUMN_VEHICLE_DATE + " text not null, "
+                    + COLUMN_VEHICLE_ODO + " text not null, "
+                    + COLUMN_VEHICLE_EVENT + " text not null, "
+                    + COLUMN_VEHICLE_PRICE + " text not null, "
+                    + COLUMN_VEHICLE_COMMENT + " text not null,"
+                    + COLUMN_ICON + " text not null default '0'"
+                    + ");");
+
+
+            String QUERY_FROM = String.format("SELECT %s, %s, %s, %s, %s, %s, %s FROM %s;",
+                    COLUMN_VEHICLE_REFID, COLUMN_ICON, COLUMN_VEHICLE_DATE, COLUMN_VEHICLE_ODO,
+                    COLUMN_VEHICLE_EVENT, COLUMN_VEHICLE_PRICE, COLUMN_VEHICLE_COMMENT,
+                    TABLE_VEHICLE);
+
+            Cursor cursorFrom = db.rawQuery(QUERY_FROM, null);
+            try {
+                if (cursorFrom.moveToFirst()) {
+                    do {
+                        ContentValues contentValues = new ContentValues();
+                        contentValues.put(COLUMN_VEHICLE_REFID, cursorFrom.getString(cursorFrom.getColumnIndex(COLUMN_VEHICLE_REFID)));
+                        contentValues.put(COLUMN_ICON, cursorFrom.getString(cursorFrom.getColumnIndex(COLUMN_ICON)));
+                        contentValues.put(COLUMN_VEHICLE_DATE, cursorFrom.getString(cursorFrom.getColumnIndex(COLUMN_VEHICLE_DATE)));
+                        contentValues.put(COLUMN_VEHICLE_ODO, cursorFrom.getColumnIndex(COLUMN_VEHICLE_ODO));
+                        contentValues.put(COLUMN_VEHICLE_EVENT, cursorFrom.getString(cursorFrom.getColumnIndex(COLUMN_VEHICLE_EVENT)));
+                        contentValues.put(COLUMN_VEHICLE_PRICE, cursorFrom.getString(cursorFrom.getColumnIndex(COLUMN_VEHICLE_PRICE)));
+                        contentValues.put(COLUMN_VEHICLE_COMMENT, cursorFrom.getString(cursorFrom.getColumnIndex(COLUMN_VEHICLE_COMMENT)));
+
+                        db.insertOrThrow(TABLE_VEHICLE_NEW, null, contentValues);
+
+                    } while (cursorFrom.moveToNext());
+                    db.execSQL("DROP TABLE IF EXISTS " + TABLE_VEHICLE);
+                    db.execSQL("ALTER TABLE " + TABLE_VEHICLE_NEW + " RENAME TO " + TABLE_VEHICLE);
+
+                    db.setTransactionSuccessful();
+                    Log.i(TAG, "Much success ben haz");
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                if (!cursorFrom.isClosed()) {
+                    cursorFrom.close();
+                    db.endTransaction();
+                    //db.close();
+                }
+            }
+        }*/
     }
 }
